requirements.txt
beautifulsoup4==4.13.4
certifi==2025.7.9
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
idna==3.10
mccabe==0.7.0
mypy_extensions==1.1.0
packaging==25.0
pathspec==0.12.1
pip==25.1.1
platformdirs==4.3.8
pycodestyle==2.14.0
pyflakes==3.4.0
python-dotenv==1.1.1
requests==2.31.0
selenium>=4.0.0
soupsieve==2.7
typing_extensions==4.14.1
urllib3==2.5.0
